Add ServiceData integration tests and TestInfo mock helper
Add 8 integration tests for ServiceData/CurrentServiceData covering:
info returns AccountInfo, info returns None for missing service,
read returns data, read returns None, read auto-expands buffer,
write returns None for new key, write returns previous length, and
read/write roundtrip.

Also fixes: panic return type (void, not never), OptionalN.none for
value types (use <T>0 instead of changetype), buf.length comparison
cast.

+++ sdk/jam/service-data.ts
@@ -43,7 +43,7 @@ export class ServiceData {
   /** Get account info for this service. Returns None if the service does not exist. */
   info(): Optional<AccountInfo> {
     // Ensure buffer is large enough for the 96-byte info structure.
-    if (this.buf.length < ACCOUNT_INFO_SIZE) {
+    if (u32(this.buf.length) < ACCOUNT_INFO_SIZE) {
       this.buf = new Uint8Array(ACCOUNT_INFO_SIZE);
     }
     const result = info(this.serviceId, u32(this.buf.dataStart), 0, ACCOUNT_INFO_SIZE);
